Screen tests won't start fixed, in test routing don't work
