+++ package.json
@@ -42,7 +42,7 @@
     "test:unit": "cross-env ELECTRON=1 yarn run jest --coverage \"test/unit\"",
     "test:e2e": "cross-env ELECTRON=1 yarn run jest \"test/e2e\"",
     "test:components": "yarn run jest --ci \"test/components\"",
-    "test:screens": "yarn run jest \"test/screens\" --runInBand",
+    "test:screens": "cross-env ELECTRON=1 yarn run jest --runInBand \"test/screens\"",
     "jest": "jest --verbose --colors",
     "pack": "yarn run build && electron-builder --dir",
     "dist": "yarn run build && electron-builder",
